return `debug` to `node` crate
