--- crates/core/src/node/tests/mod.rs
@@ -1,1 +1,0 @@
-mod debug;
